// src/app/questionupload/model/parentquestion.model.ts
export interface ParentquestionModel {
    parentQuestionId: string;
    questionDescription: string;
    questionHeader: string;
    questionNote: string;
    questionCategory: string;
    questionComplexityLevel: string;
    classId: number;
    questionDescriptionImage: File;
    scannedQuestionImage: File;
}